feat: useGetProject
