+++ web-marketplace/src/components/organisms/ChooseCreditsForm/ChooseCreditsForm.utils.ts
@@ -1,10 +1,10 @@
-import { SellOrderInfo } from '@regen-network/api/lib/generated/regen/ecocredit/marketplace/v1/query';
-
 import { Currency } from 'web-components/src/components/DenomIconWithCurrency/DenomIconWithCurrency.constants';
+
+import { UISellOrderInfo } from 'pages/Projects/AllProjects/AllProjects.types';
 
 export function getFilteredCryptoSellOrders(
   currency: Currency,
-  groupCryptoSellOrders: Partial<Record<string, Array<SellOrderInfo>>>,
+  groupCryptoSellOrders: Partial<Record<string, Array<UISellOrderInfo>>>,
   retiring: boolean,
 ) {
   return groupCryptoSellOrders[currency]?.filter(
