+++ web-marketplace/src/components/templates/ProjectDetails/ProjectDetails.tsx
@@ -67,6 +67,7 @@ import { NotFoundPage } from '../../../pages/NotFound/NotFound';
 import { GettingStartedResourcesSection } from '../../molecules';
 import { ProjectTopSection } from '../../organisms';
 import useGeojson from './hooks/useGeojson';
+import { useGetProject } from './hooks/useGetProject';
 import useSeo from './hooks/useSeo';
 import { useSortedDocuments } from './hooks/useSortedDocuments';
 import { useStakeholders } from './hooks/useStakeholders';
@@ -136,53 +137,19 @@ function ProjectDetails(): JSX.Element {
   const [isBuyFlowStarted, setIsBuyFlowStarted] = useState(false);
   const [isSellFlowStarted, setIsSellFlowStarted] = useState(false);
 
-  // first, check if projectId is an off-chain project handle (for legacy projects like "wilmot")
-  // or an chain project id
-  // or and off-chain project with an UUID
-  const isOnChainId = getIsOnChainId(projectId);
-  const isOffChainUuid = getIsUuid(projectId);
-
-  const { data: sanityProjectData } = useQuery(
-    getProjectByIdQuery({
-      id: projectId as string,
-      sanityClient,
-      enabled: !!sanityClient && !!projectId,
-    }),
-  );
-
-  // if projectId is slug, query project by slug
-  const { data: projectBySlug, isInitialLoading: loadingProjectBySlug } =
-    useQuery(
-      getProjectBySlugQuery({
-        client: graphqlClient,
-        slug: projectId as string,
-        enabled: !!projectId && !isOnChainId && !isOffChainUuid,
-      }),
-    );
-
-  // else fetch project by onChainId
   const {
-    data: projectByOnChainId,
-    isInitialLoading: loadingProjectByOnChainId,
-  } = useQuery(
-    getProjectByOnChainIdQuery({
-      client: graphqlClient,
-      enabled: !!projectId && !!isOnChainId,
-      onChainId: projectId as string,
-    }),
-  );
-
-  // else fetch project by uuid
-  const {
-    data: offchainProjectByIdData,
-    isInitialLoading: loadingOffchainProjectById,
-  } = useQuery(
-    getOffChainProjectByIdQuery({
-      client: graphqlClient,
-      enabled: !!projectId && !!isOffChainUuid,
-      id: projectId,
-    }),
-  );
+    sanityProject,
+    projectBySlug,
+    loadingProjectBySlug,
+    projectByOnChainId,
+    loadingProjectByOnChainId,
+    offchainProjectByIdData,
+    loadingOffchainProjectById,
+    isBuyFlowDisabled,
+    projectsWithOrderData,
+    onChainProjectId,
+    offChainProject,
+  } = useGetProject();
 
   const slug =
     offchainProjectByIdData?.data?.projectById?.slug ||
@@ -203,14 +170,6 @@ function ProjectDetails(): JSX.Element {
     }
   }, [element]);
 
-  const projectBySlugOnChainId =
-    projectBySlug?.data.projectBySlug?.onChainId ?? undefined;
-  const projectByUuidOnChainId =
-    offchainProjectByIdData?.data.projectById?.onChainId ?? undefined;
-  const onChainProjectId = isOnChainId
-    ? projectId
-    : projectBySlugOnChainId ?? projectByUuidOnChainId;
-
   const { data: projectResponse } = useQuery(
     getProjectQuery({
       request: { projectId: onChainProjectId },
@@ -230,19 +189,6 @@ function ProjectDetails(): JSX.Element {
       enabled: !!countryCode,
     }),
   );
-
-  const offChainProjectById = offchainProjectByIdData?.data.projectById;
-  const publishedOffchainProjectById = offChainProjectById?.published
-    ? offChainProjectById
-    : undefined;
-  const publishedOffchainProjectBySlug = projectBySlug?.data?.projectBySlug
-    ?.published
-    ? projectBySlug?.data?.projectBySlug
-    : undefined;
-
-  const offChainProject = isOnChainId
-    ? projectByOnChainId?.data.projectByOnChainId
-    : publishedOffchainProjectById ?? publishedOffchainProjectBySlug;
 
   /* Credit class */
 
@@ -333,10 +279,6 @@ function ProjectDetails(): JSX.Element {
     loadingProjectBySlug ||
     loadingOffchainProjectById;
 
-  const { isBuyFlowDisabled, projectsWithOrderData } = useBuySellOrderData({
-    projectId: onChainProjectId,
-  });
-
   const { credits } = useCreateSellOrderData({
     projectId: projectsWithOrderData[0]?.id,
   });
@@ -397,7 +339,6 @@ function ProjectDetails(): JSX.Element {
     setIsBuyFlowStarted,
   });
 
-  const sanityProject = sanityProjectData?.allProject?.[0];
   const projectPrefinancing = sanityProject?.projectPrefinancing;
   const isPrefinanceProject = projectPrefinancing?.isPrefinanceProject;
