+++ web-marketplace/src/graphql/sanity/ProjectById.graphql
@@ -39,5 +39,9 @@ query ProjectById($id: String!) {
       }
       supportEnables
     }
+    fiatSellOrders {
+      sellOrderId
+      usdPrice
+    }
   }
 }
